Move classes to features package
