+++ src/main/java/net/earthcomputer/clientcommands/command/arguments/MojmapPacketClassArgumentType.java
@@ -6,7 +6,7 @@ import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
-import net.earthcomputer.clientcommands.MappingsHelper;
+import net.earthcomputer.clientcommands.features.MappingsHelper;
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
 import net.fabricmc.mappingio.tree.MappingTreeView;
 import net.minecraft.commands.SharedSuggestionProvider;
